change some data type

+++ src/api/futures/public/instruments/mod.rs
@@ -50,7 +50,7 @@ pub struct Instrument {
     pub contract_value_trade_precision: Option<u32>,
     pub fee_schedule_uid: Option<String>,
     pub funding_rate_coefficient: Option<u32>,
-    pub impact_mid_size: Option<u32>,
+    pub impact_mid_size: Option<f64>,
     pub isin: Option<String>,
     pub last_trading_time: Option<String>,
     #[serde(default)]
